+++ src/main/java/net/telepathicgrunt/bumblezone/modinit/BzConfiguredFeatures.java
@@ -1,7 +1,10 @@
-package net.telepathicgrunt.bumblezone.world.features;
+package net.telepathicgrunt.bumblezone.modinit;
 
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
+import com.mojang.datafixers.util.Pair;
 import net.minecraft.block.Blocks;
+import net.minecraft.entity.EntityType;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.BuiltinRegistries;
 import net.minecraft.util.registry.MutableRegistry;
@@ -18,6 +21,7 @@ import net.telepathicgrunt.bumblezone.Bumblezone;
 import net.telepathicgrunt.bumblezone.modinit.BzBlocks;
 import net.telepathicgrunt.bumblezone.modinit.BzPlacements;
 import net.telepathicgrunt.bumblezone.modinit.BzFeatures;
+import net.telepathicgrunt.bumblezone.world.features.configs.NbtFeatureConfig;
 
 public class BzConfiguredFeatures {
     private static final SpringFeatureConfig SUGAR_WATER_SPRING_CONFIG = new SpringFeatureConfig(BzBlocks.SUGAR_WATER_FLUID.getDefaultState(), true, 4, 1, ImmutableSet.of(Blocks.HONEY_BLOCK, Blocks.HONEYCOMB_BLOCK));
@@ -26,12 +30,29 @@ public class BzConfiguredFeatures {
     public static final ConfiguredFeature<?,?> SUGAR_WATERFALL_FULL_RANGE = Feature.SPRING_FEATURE.configure(SUGAR_WATER_SPRING_CONFIG).decorate(Decorator.RANGE.configure(new RangeDecoratorConfig(40, 0, 128)).spreadHorizontally().repeat(1));
     public static final ConfiguredFeature<?,?> HONEYCOMB_CAVES = BzFeatures.HONEYCOMB_CAVES.configure(FeatureConfig.DEFAULT).decorate(Decorator.NOPE.configure(DecoratorConfig.DEFAULT));
     public static final ConfiguredFeature<?,?> HONEYCOMB_HOLE = BzFeatures.HONEYCOMB_HOLE.configure(FeatureConfig.DEFAULT).decorate(BzPlacements.HONEYCOMB_HOLE_PLACER.configure(DecoratorConfig.DEFAULT));
-    public static final ConfiguredFeature<?,?> BEE_DUNGEON = BzFeatures.BEE_DUNGEON.configure(FeatureConfig.DEFAULT).decorate(BzPlacements.BEE_DUNGEON_PLACER.configure(DecoratorConfig.DEFAULT));
-    public static final ConfiguredFeature<?,?> SPIDER_INFESTED_BEE_DUNGEON = BzFeatures.SPIDER_INFESTED_BEE_DUNGEON.configure(FeatureConfig.DEFAULT).decorate(BzPlacements.BEE_DUNGEON_PLACER.configure(DecoratorConfig.DEFAULT));
     public static final ConfiguredFeature<?,?> HONEY_CRYSTALS_COMMON = BzFeatures.HONEY_CRYSTAL_FEATURE.configure(FeatureConfig.DEFAULT).decorate(BzPlacements.RANDOM_3D_UNDERGROUND_CHUNK_PLACEMENT.configure(new CountConfig(4)));
     public static final ConfiguredFeature<?,?> HONEY_CRYSTALS_UNCOMMON = BzFeatures.HONEY_CRYSTAL_FEATURE.configure(FeatureConfig.DEFAULT).decorate(BzPlacements.RANDOM_3D_UNDERGROUND_CHUNK_PLACEMENT.configure(new CountConfig(2)));
     public static final ConfiguredFeature<?,?> HONEY_CRYSTALS_RARE = BzFeatures.HONEY_CRYSTAL_FEATURE.configure(FeatureConfig.DEFAULT).decorate(BzPlacements.RANDOM_3D_UNDERGROUND_CHUNK_PLACEMENT.configure(new CountConfig(1)));
     public static final ConfiguredFeature<?,?> CAVE_SUGAR_WATERFALL = BzFeatures.CAVE_SUGAR_WATERFALL.configure(FeatureConfig.DEFAULT).decorate(Decorator.RANGE.configure(new RangeDecoratorConfig(8, 0, 248)).spreadHorizontally().repeat(75));
+
+    public static final ConfiguredFeature<?, ?> BEE_DUNGEON = BzFeatures.BEE_DUNGEON.configure(
+            new NbtFeatureConfig(
+                    new Identifier(Bumblezone.MODID, "bee_dungeon_processors"),
+                    new Identifier("empty"),
+                    ImmutableList.of(Pair.of(EntityType.BEE, 1)),
+                    ImmutableList.of(Pair.of(new Identifier(Bumblezone.MODID, "bee_dungeon"), 1)),
+                    -4
+            )).decorate(BzPlacements.BEE_DUNGEON_PLACER.configure(DecoratorConfig.DEFAULT));
+
+    public static final ConfiguredFeature<?, ?> SPIDER_INFESTED_BEE_DUNGEON = BzFeatures.SPIDER_INFESTED_BEE_DUNGEON.configure(
+            new NbtFeatureConfig(
+                    new Identifier(Bumblezone.MODID, "spider_infested_bee_dungeon_processors"),
+                    new Identifier("empty"),
+                    ImmutableList.of(Pair.of(EntityType.BEE, 1)),
+                    ImmutableList.of(Pair.of(new Identifier(Bumblezone.MODID, "bee_dungeon"), 1)),
+                    -4
+            )).decorate(BzPlacements.BEE_DUNGEON_PLACER.configure(DecoratorConfig.DEFAULT));
+
 
     public static void registerConfiguredFeatures(){
         MutableRegistry<ConfiguredFeature<?, ?>> registry = (MutableRegistry<ConfiguredFeature<?, ?>>) BuiltinRegistries.CONFIGURED_FEATURE;